Implement Upload single image to server
