+++ OpenOversight/app/templates/submit.html
@@ -3,7 +3,7 @@
 
 <div class="container theme-showcase" role="main">
    <div class="page-header">
-        <h1>Submit Images/Videos <small>containing faces/badge numbers of uniformed police officers</small></h1>
+        <h1>Submit Image <small>containing faces/badge numbers of uniformed police officers</small></h1>
    </div>
 
    <div class="input-group input-group-lg">
